+++ packages/uniapp-request/src/types/index.ts
@@ -1,4 +1,2 @@
-import { type } from "os";
-
 export * from './config';
 export * from './request';