+++ packages/uniapp-request/src/types/request.ts
@@ -11,4 +11,6 @@ export interface RequestOptions {
     withCredentials?: boolean;
     firstIpv4?: boolean;
     retryCount?: number;
+    loading?: boolean;
+    loadingText?: string;
 };